requirements.txt
fastapi
uvicorn
streamlit
langchain
chromadb
pymongo
pypdf
python-dotenv
sentence-transformers
pytest
pydantic